this commit renames the Sort Toggle component
